To delete spam/formatter/requirements.txt
# Core pipeline
pymupdf>=1.23.0          # PDF text + image extraction
python-docx>=1.1.0       # DOCX extraction
jinja2>=3.1.0            # LaTeX template rendering
flask>=3.0.0             # Web UI

# Optional: install pix2text for better formula extraction from PDFs
# pip install pix2text
# (without it, PyMuPDF is used as fallback — works fine for text-heavy PDFs)
